+++ src/3_game_loop/a_game_logic/player_movement.c
@@ -1,36 +1,57 @@
 #include "../../../include/cub3d.h"
 
-void	handle_player_move(void)
+double	get_mvmnt_speed(void)
+{
+	double	speed;
+
+	speed = MOVEMENT_SPEED;
+	if (mlx_is_key_down(game()->mlx, MLX_KEY_LEFT_SHIFT))
+		speed *= MOVEMENT_SPRINT_MULTIPLIER;
+	return (speed);
+}
+
+t_vec2	get_movement_from_key(void)
 {
 	double	speed;
 	double	angle;
+	t_vec2	mvmnt;
 
-	speed = MOVEMENT_SPEED;
+	speed = get_mvmnt_speed();
 	angle = deg_to_rad(dir_vec_to_deg(player()->transform.rot));
-	if (mlx_is_key_down(game()->mlx, MLX_KEY_LEFT_SHIFT))
-		speed *= MOVEMENT_SPRINT_MULTIPLIER;
 	if (mlx_is_key_down(game()->mlx, MLX_KEY_W))
 	{
-		write(1, "W\n", 2);
-		player()->transform.pos.x = cos(angle) * MOVEMENT_SPEED;
-		player()->transform.pos.y = sin(angle) * MOVEMENT_SPEED;
+		mvmnt.x = cos(angle) * speed;
+		mvmnt.y = sin(angle) * speed;
 	}
 	if (mlx_is_key_down(game()->mlx, MLX_KEY_A))
 	{
-		write(1, "A\n", 2);
-		player()->transform.pos.x = sin(angle) * MOVEMENT_SPEED;
-		player()->transform.pos.y = -cos(angle) * MOVEMENT_SPEED;
+		mvmnt.x = sin(angle) * speed;
+		mvmnt.y = -cos(angle) * speed;
 	}
 	if (mlx_is_key_down(game()->mlx, MLX_KEY_S))
 	{
-		write(1, "S\n", 2);
-		player()->transform.pos.x = -cos(angle) * MOVEMENT_SPEED;
-		player()->transform.pos.y = -sin(angle) * MOVEMENT_SPEED;
+		mvmnt.x = -cos(angle) * speed;
+		mvmnt.y = -sin(angle) * speed;
 	}
 	if (mlx_is_key_down(game()->mlx, MLX_KEY_D))
 	{
-		write(1, "D\n", 2);
-		player()->transform.pos.x = -sin(angle) * MOVEMENT_SPEED;
-		player()->transform.pos.y = cos(angle) * MOVEMENT_SPEED;
+		mvmnt.x = -sin(angle) * speed;
+		mvmnt.y = cos(angle) * speed;
 	}
+	return (mvmnt);
+}
+
+void	handle_player_move(void)
+{
+	t_vec2	original_pos;
+	t_vec2	movement;
+
+	original_pos = player()->transform.pos;
+	movement = get_movement_from_key();
+	player()->transform.pos.x += movement.x;
+	if (*game()->input_data->map[(int)player()->transform.pos.y][(int)player()->transform.pos.x] != FLOOR)
+		player()->transform.pos.x = original_pos.x;
+	player()->transform.pos.y += movement.y;
+	if (*game()->input_data->map[(int)player()->transform.pos.y][(int)player()->transform.pos.x] != FLOOR)
+		player()->transform.pos.y = original_pos.y;
 }